+++ R/clean_indicators.R
@@ -23,7 +23,7 @@ clean_indicators <- function(df){
       indicator = indicator_code %>%
         stringr::str_extract("[^\\.]+") %>%
         dplyr::recode("VL_SUPPRESSED" = "VL_SUPPRESSION_SUBNAT"),
-      indicator = ifelse(indicator_code == "PrEP_CT.TestResult", "PrEP_CT.TestResult", indicator),
+      indicator = ifelse(indicator_code == "PrEP_CT.TestResult", indicator_code, indicator),
       numeratordenom = ifelse(stringr::str_detect(indicator_code, "\\.D\\.|\\.D$"), "D", "N"),
       statushiv = stringr::str_extract(indicator_code, "(Neg|Pos|Unk)$"),
       statushiv = ifelse(indicator == "PrEP_CT.TestResult", "Neg", statushiv),
@@ -46,11 +46,11 @@ clean_indicators <- function(df){
   #convert external modalities
   df <- convert_mods(df)
 
-  #add HTS_TST_POS as an indicator
-  df <- df %>%
-    dplyr::bind_rows(df %>%
-                       dplyr::filter(indicator == "HTS_TST" & statushiv == "Positive") %>%
-                       dplyr::mutate(indicator = "HTS_TST_POS"))
+  #add calculated indicators (HTS_TST_POS, OVC_HIVSTAT_D, PMTCT_STAT_POS, PMTCT_ART)
+  df <- calculate_inds(df)
+
+  #add trendscoarse
+  df <- align_agecoarse(df)
 
   #move keypop to otherdisagg
   df <- df %>%